requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
factory-boy==3.2.0
Faker==5.1.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.2
whitenoise==5.2.0
